requirements.txt
pandas>=2.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
Flask>=3.0.0
Werkzeug>=3.0.1
gunicorn>=21.2.0
